Take storage out of the FluidPort
MTK gave a new warning:

> Warning: x contains 1 flow variables, yet 2 regular
(non-flow, non-stream, non-input, non-output) variables.
This could lead to imbalanced model that are difficult to debug.
Consider marking some of the regular variables as input/output variables.

Since pressure/head is a potential for flow, it is best to keep that in the FluidPort as is common in Modelica. Storage can be seen as a measured output.