start refactor to allow networked systems
Not yet complete, but this code works now for single LSWs, so a good moment to pause.

This reintroduces the connectors, and pulls the exchanges such a watermanagement and weir outflow out of the LSW component, making it generic to both P and V types.

The V type now also get a level lookup. This is the level of the most downstream weir area. It's easier to interpret levels than volumes in the output.