add concentration to the flow
Besides adding concentration as a stream to the connector, this also is a refactoring of construction.jl

Before we created nested systems, lsw_sys with the lsw and related systems together. This seemed to create namespacing issues when connecting components from different levels. Now this intermediate system no longer exists. Any system that belongs to an LSW has the LSW ID in the name. For users this naming scheme still needs to be finished, since currently we only support one user.