use simulation starttime as t = 0, not 1970
Internally we need to represent time as a floating point number. That means we need a method to convert from DateTime to this float and vice-versa. So far we used Unix time, seconds since 1970. This was convenient at first since there was a global datum.

There were several downsides to using unix time however. First of all, the floats were hard to read when debugging: `datetime2unix(now()) == 1.677761097515e9`. The precision is enough for Float64 (`eps(ans)
2.384185791015625e-7`), but if we want to move to Float32 it could give issues, especially when moving further away from 1970.

So instead this maps the starttime to 0.0. This will have more precision, and to avoid hard to debug accuracy errors we error out if `eps(t_end)` is higher than 1 hour. We need the `config.starttime` to convert, but usually have Config available anyway. I put the Config in the Register struct as well to be able to map times after running a simulation as well.

Note: I considered two other options. Option 1 was to always run from 0 to 1, but not having a fixed time unit is inconvenient, and not convenient for the BMI for instance. I also considered using days since starttime, but then you need to do `du .*= 86400` to get `du` right. which doesn't feel great.

BMI also states most models start time at 0.0: https://bmi.readthedocs.io/en/stable/#get-start-time