allow setting some solver options for some solvers
Fixes #118.
This removes the `saveat` key from the config root, and moves it to the new `solver` section, together with `algorithm` and some other options. The `saveat` Union type is extended, since in TOML [] becomes a Union{}[] which cannot convert to Float64[].

The algorithms were chosen to cover some different approaches. This switches the default to QNDF, but for that to work we need to set `autodiff=false`, probably until #40 is fixed.

```TOML
[solver]
autodiff = false
```

This still needs docs and proper Python support, I hardcoded `autodiff=false` in the TOML writing for now.