add Pump (#127)
Fixes #21. Except for the devdocs (#126) and unit tests. Since there are
no proper unit tests for node types yet, I prefer to fix #116 first
before addressing that.

A pump is added to the basic model which runs in the docs, and based on
the output from there it seems to work ok, stopping to pump when the
upstream basin is empty.

---------

Co-authored-by: Hofer-Julian <30049909+Hofer-Julian@users.noreply.github.com>