use regular Dict, remove Dictionaries (#128)
From the Dictionaries README:

> The three main difference to Dict are that it preserves the order of
elements, it iterates much faster, and it iterates values rather than
key-value pairs.

In the current code we have only one `Dictionary`, the `u_index` (basin
node ID to state index), and we don't iterate over it. So it seems to me
the advantages of Dictionaries don't apply anymore, which means we can
drop a dependency.