prepare model input from python and add tests (#149)
This tackles the larger part of #116.

This create the example models under `./data/`:

```
python python/ribasim/tests/conftest.py
```

This runs before the Julia tests in GitHub Actions, and on TeamCity
before building binaries. That does mean that a functioning
`ribasim-python` is needed to run the Julia tests.

If the example models change, the script needs to be re-run manually at
the moment. It would probably be better to run the script from Julia at
the beginning of the tests to ensure they are always up to date. Because
normally Julia packages should be testable with just `Pkg.test`, without
any other steps. For that to be convenient locally as well we should
probably look a bit into
[PythonCall.jl](https://github.com/cjdoris/PythonCall.jl) and perhaps
set up a way that it can re-use the same conda environment.