simplify: move to StructArrays
Starting with the tabulated rating curve. But the intention is to get rid of DataFrames, and use StructArrays. For tables from GeoPackage, these are copied to columntables (a NamedTuple) and then wrapped in a StructArray. For Arrow tables we can do the same, except the StructArray will be backed by memory mapped Arrow data.