fix output IDs (#172)
The `sort` is important here since we use a regular unordered Dict. This
resulted in the wrong `node_id` in the header of the basin output.

Also adds `edge_id` to the flow output, as intended in #163.

No tests right now since we don't have any tests on the output tables
yet. I createad #171 for that.