don't use geopandas for nonspatial table IO
Only the node and edge table support geometries. This copies the input_base IO functions to the node and edge classes, to special case them. These functions are the same for node and edge, so there is a bit of duplication there.

In input_base we use the sqlite3 stdlib and pandas to read and write. I use `pd.read_sql_query` instead of `pd.read_sql_table` because that way it works with `sqlite3` stdlib instead of needing SQLAlchemy. Pandas refers to the sqlite3 functionality as legacy, but I thought it's not worth another dependency.

Also fixes #173, by deleting old geopackage files if they exist.