update forcings in callback
Before we loaded the whole timeseries for Basin into memory and put a LinearInterpolation onto it. Then during each solver iteration the interpolation object was queried with the current time t.

A downside of this approach was that there are no guarantees that your forcing ends up in your model. The solver could possibly take a large timestep and completely miss a rainfall peak. To avoid this we need to stop the solver every time there is new forcing data available, and then load it into the model.

With this we can currently only do forward fill interpolation. This might be a bit harder on the solver since a precipitation peak will come as a shock. We could consider building something on top of this that provides some smoothing, though fundamentally I think this is the correct and conservative thing to do.

Some extra utility functions were added to load the initial forcing parameters. These are generic and will be used for time tables of other nodes as well.

The transient model state deviation is small and expected as explained above. The TabulatedRatingCurve change is large. I recalculated the water balance there, and the new one is correct. It looks like rain fall on the downstream basin before, but that was not intended.