use get_level which supports both Basin and LevelBoundary
Before we had u_index which gave the Basin index from an ID. This was in the Connectivity struct. Now it is removed, but we added the node_id field to the basin struct to support the same functions.

We move back to Dictionaries since this conveniently allows us to calculate the ID index once during initialization. This index can then be re-used to look up the other fields, such as current_level.