get the ManningResistance running
Leave out the slope_unit_length optimization for now, I suspect it doesn't help too much, plus it could lead to errors in the future if we update the slope but this is not recomputed.

The actual resistance seems very small, also if I decrease the width and increase the length. This still needs verification.