remove unused dirs; run black and ruff on whole repo (#233)
The `pre` folder was used to derive national schematisations. This was
no longer in use, and in the future this will be done using
ribasim-python.

The `modflow` folder contained code for coupling with metaswap and
modflow. These couplings will be implemented in imod-coupler using
libribasim.

This also runs black and ruff on the whole repo, such that
examples.ipynb and qgis/ also get that treatment, towards #231.