calculate storage from Basin / profile (#280)
Fixes #225

This removes the need to provide the storage for a profile. We can
calculate it ourselves by integrating over A(h) to get A(S) and h(S).

First this updates the test models to a correct storage in the profile.
Then we calculate the profile and get the same results. Then we remove
the storage from the schemas and docs.

I hoped this would help with #80 but based on the plot I added there it
doesn't seem to. This does not yet add all the validation and
extrapolation for lookup tables as described in #279, that can be done
separately.