More accurate interpolation (#314)
I have implemented the physical interpolation (as proposed in
https://github.com/Deltares/Ribasim/issues/225). I have not checked
whether this implementation is exactly what we want for extrapolation
(https://github.com/Deltares/Ribasim/issues/279), but this
implementation assumes that the storage is 0 at the lowest level and
does linear extrapolation of the area as a function of the level based
on the last 2 values.

This does not solve the problem of high oscillating flows produced by
the ManningResistance node as hoped
(https://github.com/Deltares/Ribasim/issues/80). I think the problem
there is that the calculated flows are way to large, and therefore the
equilibrium is overshooted at every time step.

I cannot wrap my head around the current ManningResistance
implementation. I understand that flow occurs due to a difference in
level, but I do not understand why that means that all the water comes
in motion, not just the bit at the top (which is why I find
LinearResistance more intuitive). Is anyone familiar with a
ManningResistance being used for flow that can go both directions, or is
it only used for flow that goes one direction due to gravity? In that
last case I understand better why all water is in motion.

Also, the current implementation assumes the presence of a lot of water
in the channel itself, which fluctuates with the level in the upstream
basin with no regard for (local) water preservation (by which I do not
mean that the ManningResistance node is not preserving but it is less
physical).

---------

Co-authored-by: Bart de Koning <bart123-4@hotmail.com>
Co-authored-by: Martijn Visser <mgvisser@gmail.com>