add solver setting "adaptive" (#450)
If an algorithm supports adaptive timestepping, `dt` only means the
initial timestep. We need to set adaptive to false to be able to run
with fixed timesteps.

In the future we might want to consider setting this to false by
default, but for not let's not flip that switch, and follow SciML
default adaptive=true.