ribasim_cli: only return 0 on successful run (#524)
Before we returned 0 if the model did not crash. Now it returns a 1 if
there is for instance a convergence error.

This also adds a println for the SciML return code. On successful runs
this will print "Success", with other options documented here:
https://docs.sciml.ai/SciMLBase/stable/interfaces/Solutions/#Specific-Return-Codes

Perhaps in the future this println can be integrated into the logging
system better, but for now we need a way to always communicate this info
to cli users, since otherwise they don't know about convergence issues.

Also updates the manifests.