Use FixedSizeDiffCache for flows (#581)
When profiling runs with the default `autodiff=true`, this line was
responsible for 35% of the time and almost all allocations:


https://github.com/Deltares/Ribasim/blob/b3eb044a722d1655c5465bafe50951b75fe960d6/core/src/solve.jl#L1002

With this PR that drops down to 0%.

`connectivity.flow` is a sparse matrix, but the DiffCache does not seem
to like sparse matrixes. The `dual_du` field was a dense vector of
length n x n x cache_size, and the `get_tmp` call led to further
allocations trying to restructure the sparse matrix from the vector.
Luckily there is the FixedSizeDiffCache that helps here:
https://docs.sciml.ai/PreallocationTools/stable/#FixedSizeDiffCache

This retains the sparsity in the dual, and returns a `ReinterpretArray`
from `get_tmp` during autodiff. To avoid materializing this
reinterpretarray I needed to additionally fill the parent array with
zeros rather than the array itself.

There is another unrelated performance fix here, and that is to
concretely type the Parameter struct, by adding type parameters from its
fields. Otherwise you have situations like

```julia
struct A
    a::Vector
end
```

where the compiler doesn't know the element type of the Vector, so it
can perform less optimizations. The solution:

```julia
struct A{T}
    a::Vector{T}
end
```

Finally I consistently added AbstractVector/Matrix argument type
annotations to ensure the ReinterpretArray could enter everywhere. And I
renamed the functions to formulate flows to `formulate_flow`, to make it
easier to separate them from the other `formulate!` methods.