Only gravity driven `Outlet` flow (#549)
Fixes https://github.com/Deltares/Ribasim/issues/548.

Currently the outlet switches on and off in a discrete way due to a
switch in sign of the level difference. This can result in the outlet
rapidly turning on and off:


![flipperende_outlet](https://github.com/Deltares/Ribasim/assets/74617371/d834ffa9-ac48-4fb1-99c5-732369f43c35)

To solve this problem, a reduction factor should be applied to the
outlet flow, based on the level difference. This reduction factor will
have to be taken into account in the implementation of the analytical
Jacobian (with an unsatisfying effort to improvement ratio). This I
think will be a recurring problem until we are able to switch to an AD
Jacobian.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>