add runstats utility (#611)
This addresses a part of #462. It is a way to run and log performance
related characteristics into a table. That table can then be used to do
further analysis on what the effects of e.g. code changes or solver
settings are.

This does not yet hook it up to anything, so there is the possibility
for bitrot. But I've had this locally for a while, and I wanted to share
it. Even if this doesn't run in TeamCity every week it can be useful for
local analysis.

There are many columns, and one row per run, so a small part of the
table looks like:


![image](https://github.com/Deltares/Ribasim/assets/4471859/eae20bce-5b09-47a3-b6e8-d81e7dfeb114)

Columns are about solver settings, host machine, julia version, git
commit, timing and GC.