start using isoutofdomain
This rejects any steps that lead to a negative storage. See also:

https://docs.sciml.ai/DiffEqDocs/stable/basics/faq/#My-ODE-goes-negative-but-should-stay-positive,-what-tools-can-help?

When using adaptive solvers this will cause the dt to go down. If it is not adaptive and implicit, it will fail:

```
Warning: Newton steps could not converge and algorithm is not adaptive. Use a lower dt.
```

Both options seem desired behavior.

I managed to still get negative storages with an explicit solver with a large fixed dt, and a flux that empties a basin in well under a timestep. But adaptive is the default and should generally be used.

The `du` correction in `water_balance!` is also removed as it should normally not be needed anymore. And if it does, I'd prefer to not distort the water balance.

One alternative is to use the PositiveDomain callback, but this only works for adaptive=false, and is not easily adapted to enforce only the storage part of the state to be positive.