refactor writing results
Instead of separate functions per output file that both create the table and write it to an Arrow file, this splits it up to a function that creates the in-memory table as a NamedTuple of Vectors, and a shared function that writes the Arrow file.

The functions to create the in-memory functions will also be useful for testing the output.