Docs: update download instructions and iMOD QGIS install (#664)
Since @JoerivanEngelen released a new version of the iMOD plugin, we can
simply tell users to download from the plugin repository. This updates
and extends the QGIS docs a little, linking to the iMOD documentation as
well.

This also updates the download sections on the documentation homepage
and the readme. The documentation for separate components now refers to
this section. For the release binaries I now link to the GitHub release
assets rather than the Amazon S3 link.

This also adds more badges, because they are so nice and green:


![image](https://github.com/Deltares/Ribasim/assets/4471859/7d356be8-2adb-49b9-8d9e-03a6220547c0)