rename model files (#690)
Fixes a large part of #638. Not entirely, I listed two follow-ups that
are probably best done after #630 in that issue.
This also removes the model name from the model, as before the only
thing it did was name the TOML and GeoPackage.
That meant I had to adapt `ribasim_testmodels` a bit to expose a
`constructors` dict which maps the test model names to the constructor
functions.


https://github.com/Deltares/Ribasim/compare/names?expand=1#diff-b223bf31f9400e170f97d95fa1d8b22393e5d37cfd810a4a5aa7bf480915ac52

This doesn't just change the file names but also updates the
documentation to stop talking about GeoPackage so much and consistently
refer to the database. Also QGIS panel names are updated, function,
class and variable names.

TOML section names are updated, so this is breaking. Specifically, the
`geopackage` key is now called `database`, and the `[output]` section is
now called `[results]`. The default output folder is now also called
results.