Use a TimeZones.jl fork for now (#698)
Alternative to #687, which wasn't working well with PackageCompiler due
to long paths still, leading to broken builds.

This switches to a fork. A PR to merge this upstream has already been
submitted: https://github.com/JuliaTime/TimeZones.jl/pull/450

Clearly this is temporary, but since we use Manifests this should work
ok.