Update to python 3.12, pandas 2.1.2
Fixes #654
Fixes #645

I ran `pixi add pandas=2.1.2 python=3.12 pyogrio=0.7.2` using pixi v0.5.0 so as not to update the lockfile format.

The new pyogrio release is needed to be able to use Python 3.12. We only update the Python version in the Pixi lockfile and on CI, this does not change our minimum supported Python version.

The pandas update to 2.1.2 fixes a regression present in 2.1.0 and 2.1.1, no further changes were needed in our code.