Fix FractionalFlow sum to 1 validation with control states
The initial fractions come from whatever happens to be the last control state per node, and therefore can trigger false positive validation errors.

Now we validate this per control state. To make that work I had to add the missing control state to the `control_mapping`. For simplicity I coalesce a missing `control_state` to the empty string. I wonder if that is the right approach though, since for e.g. name we default to the empty string rather than support missing. We should probably settle on either for all string columns.