Share manifests (#740)
One frequent reason that `libribasim` and `ribasim_cli` builds break is
because we add a dependency to the core, and we forget to resolve the
Manifests of `libribasim` and `ribasim_cli`. (e.g. #484).

This adds an entry like `manifest = "../Manifest.toml"` to the `core`,
`libribasim` and `ribasim_cli` Project.toml, which all point to the new
`root` Manifest. With this they still have their own Project, but a
shared Manifest. This option exists to make it easier to work with
multiple related projects in one monorepo.

~~As a follow-up we could consider also doing~~ We also do this for the
remaining two projects, `docs` and `create_binaries`. `create_binaries`
doesn't depend on Ribasim, so there is not as much benefit. `docs` does,
currently not via Project.toml, but via adding it to the LOAD_PATH. It
would be cleaner to also use a shared Manifest there. One thing to check
if the unmaintained docs dependencies don't hold back Ribasim
dependencies, see #621 (EDIT: they don't).

~~This is on top of #738, but could be separated out if needed.~~

Also simplifies the docs, and has a not-directly-related commit
8403bef7b117ca0b8eadfd52528a4adfe518ada9 that runs CompatHelper for all
projects, not just core.

For reference, the root project is populated like this: `dev ./core
./docs build/create_binaries build/libribasim build/ribasim_cli`.