update lockfile with pixi 0.7 (#787)
I noticed with #731 we went back from Python 3.12 to 3.11. This updates
all dependencies, including Python back to 3.12.

I did this using pixi 0.7.0 which just came out:
https://github.com/prefix-dev/pixi/releases/tag/v0.7.0

The lockfile format seems the same, so I'm hoping this works on TeamCity
with pixi 0.6.0, but I triggered tests to make sure. There is a new top
level `version: 2` entry on top of the lockfile, so it looks like they
started versioning it.