Support toml in QGIS plugin (#781)
Fixes part of #629 (the results are not automatically associated yet).

The Open button now supports toml instead of gpkg. The toml file checks
for the "database" entry and loads that file as it used to do.
The New button now also creates a toml file with the database entry
pointing to the gpkg that is also created like before. And the start and
end dates are 01-01-2020 until 01-01-2030.

An option for another PR would be to add the start and end dates as
fields in the main widget so that they can be adjusted. But that would
mean that we need to keep the open toml file in memory and set the start
and end date, and flush the file after those changes.

Also, error handling is not fancy yet. If the "database" entry is
missing from the toml file, you will get a python KeyError

The folder `qgis` was renamed to `ribasim_qgis` for pylance to handle
the imports better.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>