Opt out of Aqua 0.8 persistent_tasks test (#791)
The release of Aqua 0.8 broke CI. That won't happen again since on main
now we declare the compat. It seems that we don't work with Aqua 0.8
yet, so we stick to 0.7 for now. The error is:

```
┌ Error: Unexpected error: /tmp/jl_LSHYf656FI/done.log was not created, but precompilation exited
└ @ Aqua ~/.julia/packages/Aqua/rTj6Y/src/persistent_tasks.jl:159
  Activating project at `/tmp/jl_b4guN0`
Persistent tasks: Test Failed at /home/runner/.julia/packages/Aqua/rTj6Y/src/persistent_tasks.jl:80
  Expression: !(has_persistent_tasks(package; kwargs...))
```

from
https://github.com/Deltares/Ribasim/actions/runs/6890693118/job/18744259890

We can opt-out of this test with `persistent_tasks=false`, see also
https://juliatesting.github.io/Aqua.jl/dev/#Aqua.test_persistent_tasks-Tuple%7BBase.PkgId%7D
and
https://juliatesting.github.io/Aqua.jl/dev/#Aqua.find_persistent_tasks_deps-Tuple%7BBase.PkgId%7D.

Created #792 to track enabling this.