Drop Python 3.9 support (#795)
Fixes #653. The exception here is QGIS. Since it still comes bundled
with Python 3.9 we cannot really drop it there.

Most of the changes here are from ruff:
https://docs.astral.sh/ruff/settings/#target-version
If you run it on a single package it picks up `requires-python` from
`pyproject.toml`, but that doesn't seem to be picked up when running it
from the repo root. Therefore it always defaulted to the ruff default
target of Python 3.8. I currently set it to 3.9 for QGIS. Though setting
it to 3.10 doesn't give extra ruff errors or update hints.