WIP Start a QgsApplication to load the plugin
The pluginPath seems to be pointing in the wrong folder.
It would be best if we make our own folder in the .pixi folder
where we put our installed plugin, and only this one.
Can we set a target with pip to a different folder?
Then we could use QgsApplication.setPluginPath() to point to that folder
for the tests.