fix documentation CI (#827)
Currently the documentation job is broken:
https://github.com/Deltares/Ribasim/actions/runs/6980836480/job/18996836760#step:10:101

I looked into it and saw that the reason for this is that we serialize
our new NodeID to Arrow instead of Int. This fixes that.

I also added extra unit tests for NodeID, and expanded the return type
annotation for all the tables we generate. To convert NodeID to Int, I
prefer `Int(id)` over `id.value` for clarity, so changing that too.

There is one 4.5 allocated flow that I set to 4.0, but I get the feeling
that this is random. @SouthEndMusic have you seen that too? If so,
perhaps we can slightly alter the test allocation problem such that it
prefers either 4.0 or 4.5.

Lastly, this updates pixi.lock, since I saw that our Python version was
different on linux only (3.10 instead of 3.11). 3.12 is not yet possible
since we need a new qgis build for that first, but people are working on
that already.