fixing mypy issues for ribasim_qgis plugin (#823)
Fixes #819 

Add mypy correctly to vs code and pixi. Required to add the mypy.ini
file to the command line in vs code, also required to specify which mypy
to use and which paths to look for.

The mypy.ini required some excludes for the qgis modules.

Also fixes a bug in the nodes widget. I broke it during #826. See
nodes_widget.py:62.

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>