Use custom iterators to speed up `inflow_ids` and friends (#830)
@Huite mentioned the performance of medium-sized models regressed
recently, so I looked into it a bit, doing some profiling. I saw a lot
of time was spent on finding the in- and outneighbors of flow. This
changed quite a bit in #807, and whilst the API is a lot nicer,
unfortunately there was also a performance regression. The profile
showed allocations. This PR removes those allocations, leading to a 2x
speedup for a test model, from 16s to 8s. Pre-#807 this ran in 3s
however, so there is still a considerable gap to close. I don't really
want to revert #807, though perhaps for flow connections we need another
data structure, or some other way to speed up.

This PR changes from using a filtered array comprehension, which creates
a `Vector{NodeID}`, to a custom iterator `InNeighbors` which iterates
over the filtered `inneighbors`. This avoids allocations but otherwise
returns the same result.