subtract allocated water from the source (#846)
I noticed while preparing a demo that allocated water was subtracted
from capacities, but not from the source. This mattered for my model
where the edges have infinite capacities. It would allocate more water
than came in when using multiple priorities.

Separated out of #842. Testing this path will be covered by the test
model from
https://github.com/Deltares/Ribasim/pull/842#issuecomment-1830045720
that we will add.