Add reduction factor to LinearResistance (#871)
Under most scenarios this shouldn't be necessary. If Basin A and B are
connected via a LinearResistance, and the bottom of A is above the water
level of B, then it is. Without a smooth landing from a reduction
factor, the storage of A shoots to a negative value, which is rejected,
leading to convergence issues.

Since we kept reusing the same pattern for calculating the reduction
factor, I factored it out to a new function `low_storage_factor`.