Switch to different TimeZones branch (#880)
This branch has an updated version number, 1.14. The benefit of this is
if you update all packages, it will not switch back to the released
version since the fork has a higher version number. That means that if
we want to update the Manifest.toml, just running `update` in the root
project is enough.