Install julia via pixi (#878)
One step less in installing your development environment. Juliaup is now
installed via pixi and the julia executable itself with the supported
version is installed in `<project>/juliaup`. The packages that we use
within julia are still shared, as they are not precompiled in
`~/.julia`.

This enables us to build our packages without using the pre-installed
versions of julia on our Teamcity build agents.

Julia is only installed once via pixi on Github Actions, it then
performs the tests over `pixi run test-ribasim-core-cov`.

Part of #93

---------

Co-authored-by: Martijn Visser <mgvisser@gmail.com>