Cleanup top-level directory
There are currently many things in the top-level directory. Most of them need to be there, but the env_setup scripts can be moved. So I did exactly that.

Note that this means that the juliaup folder will also be moved to utils. I like that to be honest, but it could be adapted.